chatpdf/requirements.txt
langchain
langchain_openai
pypdf
chromadb
langchain-chroma
tiktoken
python-dotenv